+++ src/main/java/org/itsallcode/jdbc/SimpleStatement.java
@@ -23,7 +23,7 @@ public class SimpleStatement implements AutoCloseable {
     <T> SimpleResultSet<T> executeQuery(final String sql, final ContextRowMapper<T> rowMapper) {
         final ResultSet resultSet = doExecuteQuery(sql);
         final ResultSet convertingResultSet = ConvertingResultSet.create(dialect, resultSet);
-        return new SimpleResultSet<>(context, convertingResultSet, rowMapper);
+        return new SimpleResultSet<>(context, convertingResultSet, rowMapper, this);
     }
 
     private ResultSet doExecuteQuery(final String sql) {